requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
numpy==2.3.3
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
requests==2.33.0
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
curl-cffi>=0.8.0
playwright